just uploading it all :(
Because of this, the filepath for all the CCD lab images will be slightly different - it'll be "./fits/...", rather than "./480_grpZ/...". Also the data reduction notebook assumes the Wolf1346 directory is in the current working directory. Also, for a bit of direction: band_gap_etc., defaced_gain, and linearity all pertain to the CCD lab, Data_Reduction, as one might expect, is for the data reduction lab.